+++ kompot/cli/main.py
@@ -59,6 +59,7 @@ def main():
     from .de import add_de_parser
     from .da import add_da_parser
     from .dm import add_dm_parser
+    from .impute import add_impute_parser
 
     parser = argparse.ArgumentParser(
         prog="kompot",
@@ -89,6 +90,7 @@ def main():
     add_dm_parser(subparsers)
     add_de_parser(subparsers)
     add_da_parser(subparsers)
+    add_impute_parser(subparsers)
 
     # Parse arguments
     args = parser.parse_args()